+++ lib/ffi.ts
@@ -1,4 +1,3 @@
-import { join } from "https://deno.land/std@0.170.0/path/mod.ts";
 import * as BuildSystem from "./include/BuildSystem.h.ts";
 import * as CXCompilationDatabase from "./include/CXCompilationDatabase.h.ts";
 import * as CXDiagnostic from "./include/CXDiagnostic.h.ts";
@@ -9,6 +8,7 @@ import * as Documentation from "./include/Documentation.h.ts";
 import * as FatalErrorHandler from "./include/FatalErrorHandler.h.ts";
 import * as Index from "./include/Index.h.ts";
 import * as Rewrite from "./include/Rewrite.h.ts";
+import { tryLoadLibclang } from "./baseUtils.ts";
 
 const IMPORTS = {
   ...BuildSystem,
@@ -31,74 +31,4 @@ if (!libclangPath) {
   );
 }
 
-type ClangSymbols = typeof IMPORTS;
-
-let libclang: ReturnType<
-  typeof Deno.dlopen<ClangSymbols>
->;
-
-if (Deno.build.os === "windows") {
-  if (libclangPath.includes(".dll")) {
-    libclang = Deno.dlopen(libclangPath, IMPORTS);
-  } else {
-    libclang = Deno.dlopen(
-      join(libclangPath, "libclang.dll"),
-      IMPORTS,
-    );
-  }
-} else if (Deno.build.os === "darwin") {
-  if (libclangPath.includes(".dylib")) {
-    libclang = Deno.dlopen(libclangPath, IMPORTS);
-  } else {
-    libclang = Deno.dlopen(
-      join(libclangPath, "libclang.dylib"),
-      IMPORTS,
-    );
-  }
-} else {
-  const isFullPath = libclangPath.includes(".so");
-  if (isFullPath) {
-    // if LIBCLANG_PATH point to a so file, we try to load it directly
-    libclang = Deno.dlopen(libclangPath, IMPORTS);
-  } else {
-    // Try various known libclang shared object names.
-    let lastError: null | Error = null;
-    for (
-      const file of [
-        "libclang.so",
-        "libclang.so.20",
-        "libclang.so.20.1",
-        "libclang.so.20.1.7",
-        "libclang.so.19",
-        "libclang.so.19.1",
-        "libclang.so.19.1.7",
-        "libclang.so.18",
-        "libclang.so.18.1",
-        "libclang.so.18.1.8",
-        "libclang.so.17",
-        "libclang.so.17.0",
-        "libclang.so.17.0.6",
-        "libclang.so.16",
-        "libclang.so.16.0",
-        "libclang.so.16.0.6",
-        "libclang-14.so.1",
-        "libclang.so.14.0.6",
-        "libclang.so.14",
-        "libclang.so.13",
-      ]
-    ) {
-      const fullpath = join(libclangPath, file);
-      try {
-        libclang = Deno.dlopen(fullpath, IMPORTS);
-        break;
-      } catch (e) {
-        lastError = e as Error;
-      }
-    }
-    if (lastError && !libclang!) {
-      throw lastError;
-    }
-  }
-}
-
-export { libclang };
+export const libclang = tryLoadLibclang(libclangPath, IMPORTS);
